+++ source/patterns/@aws-solutions-constructs/aws-cloudfront-s3/test/test.cloudfront-s3.test.ts
@@ -94,7 +94,7 @@ test('check existing bucket', () => {
   });
 
   const props: CloudFrontToS3Props = {
-    existingBucketObj: existingBucket
+    existingBucketInterface: existingBucket
   };
 
   new CloudFrontToS3(stack, 'test-cloudfront-s3', props);
@@ -127,11 +127,14 @@ test('test cloudfront with custom domain names', () => {
       DefaultCacheBehavior: {
         CachePolicyId: "658327ea-f89d-4fab-a63d-7e88639e58f6",
         Compress: true,
-        LambdaFunctionAssociations: [
+        FunctionAssociations: [
           {
-            EventType: "origin-response",
-            LambdaFunctionARN: {
-              Ref: "testcloudfronts3SetHttpSecurityHeadersVersionF1C744BB"
+            EventType: "viewer-response",
+            FunctionARN: {
+              "Fn::GetAtt": [
+                "testcloudfronts3SetHttpSecurityHeaders6C5A1E69",
+                "FunctionARN"
+              ]
             }
           }
         ],
@@ -214,7 +217,7 @@ test("Test bad call with existingBucket and bucketProps", () => {
   const app = () => {
     // Helper declaration
     new CloudFrontToS3(stack, "bad-s3-args", {
-      existingBucketObj: testBucket,
+      existingBucketInterface: testBucket,
       bucketProps: {
         removalPolicy: RemovalPolicy.DESTROY
       },