+++ source/patterns/@aws-solutions-constructs/core/lib/input-validation.ts
@@ -27,6 +27,7 @@ import * as kms from "@aws-cdk/aws-kms";
 export interface VerifiedProps {
   readonly dynamoTableProps?: dynamodb.TableProps,
   readonly existingTableObj?: dynamodb.Table,
+  readonly existingTableInterface?: dynamodb.ITable,
 
   readonly existingStreamObj?: kinesis.Stream;
   readonly kinesisStreamProps?: kinesis.StreamProps,
@@ -43,6 +44,7 @@ export interface VerifiedProps {
   readonly mediaStoreContainerProps?: mediastore.CfnContainerProps;
 
   readonly existingBucketObj?: s3.Bucket,
+  readonly existingBucketInterface?: s3.IBucket,
   readonly bucketProps?: s3.BucketProps,
 
   readonly topicProps?: sns.TopicProps,
@@ -75,6 +77,11 @@ export function CheckProps(propsObject: VerifiedProps | any) {
     errorFound = true;
   }
 
+  if (propsObject.dynamoTableProps && propsObject.existingTableInterface) {
+    errorMessages += 'Error - Either provide existingTableInterface or dynamoTableProps, but not both.\n';
+    errorFound = true;
+  }
+
   if (propsObject.existingStreamObj  && propsObject.kinesisStreamProps) {
     errorMessages += 'Error - Either provide existingStreamObj or kinesisStreamProps, but not both.\n';
     errorFound = true;
@@ -102,6 +109,11 @@ export function CheckProps(propsObject: VerifiedProps | any) {
 
   if (propsObject.existingBucketObj && propsObject.bucketProps) {
     errorMessages += 'Error - Either provide bucketProps or existingBucketObj, but not both.\n';
+    errorFound = true;
+  }
+
+  if (propsObject.existingBucketInterface && propsObject.bucketProps) {
+    errorMessages += 'Error - Either provide bucketProps or existingBucketInterface, but not both.\n';
     errorFound = true;
   }