+++ source/patterns/@aws-solutions-constructs/core/lib/lambda-helper.ts
@@ -16,7 +16,7 @@ import * as iam from '@aws-cdk/aws-iam';
 import * as ec2 from "@aws-cdk/aws-ec2";
 import { DefaultLambdaFunctionProps } from './lambda-defaults';
 import * as cdk from '@aws-cdk/core';
-import { overrideProps } from './utils';
+import { overrideProps, addCfnSuppressRules } from './utils';
 import { buildSecurityGroup } from "./security-group-helper";
 
 export interface BuildLambdaFunctionProps {
@@ -138,36 +138,32 @@ export function deployLambdaFunction(scope: cdk.Construct,
 
   const cfnLambdafunction: lambda.CfnFunction = lambdafunction.node.findChild('Resource') as lambda.CfnFunction;
 
-  cfnLambdafunction.cfnOptions.metadata = {
-    cfn_nag: {
-      rules_to_suppress: [{
-        id: 'W58',
-        reason: `Lambda functions has the required permission to write CloudWatch Logs. It uses custom policy instead of arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole with tighter permissions.`
-      },
-      {
-        id: 'W89',
-        reason: `This is not a rule for the general case, just for specific use cases/industries`
-      },
-      {
-        id: 'W92',
-        reason: `Impossible for us to define the correct concurrency for clients`
-      }]
+  addCfnSuppressRules(lambdafunction, [
+    {
+      id: 'W58',
+      reason: `Lambda functions has the required permission to write CloudWatch Logs. It uses custom policy instead of arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole with tighter permissions.`
+    },
+    {
+      id: 'W89',
+      reason: `This is not a rule for the general case, just for specific use cases/industries`
+    },
+    {
+      id: 'W92',
+      reason: `Impossible for us to define the correct concurrency for clients`
     }
-  };
+  ]);
 
   if (cfnLambdafunction.tracingConfig) {
     // Find the X-Ray IAM Policy
     const cfnLambdafunctionDefPolicy = lambdafunction.role?.node.tryFindChild('DefaultPolicy')?.node.findChild('Resource') as iam.CfnPolicy;
 
     // Add the CFN NAG suppress to allow for "Resource": "*" for AWS X-Ray
-    cfnLambdafunctionDefPolicy.cfnOptions.metadata = {
-      cfn_nag: {
-        rules_to_suppress: [{
-          id: 'W12',
-          reason: `Lambda needs the following minimum required permissions to send trace data to X-Ray and access ENIs in a VPC.`
-        }]
+    addCfnSuppressRules(cfnLambdafunctionDefPolicy, [
+      {
+        id: 'W12',
+        reason: `Lambda needs the following minimum required permissions to send trace data to X-Ray and access ENIs in a VPC.`
       }
-    };
+    ]);
   }
 
   return lambdafunction;